Split automation run starter
